refactor(storage): 残りの getSettings/getAllowedUrls を SettingsRepository/urlWhitelist に移行（PBI 01 Phase2-5）
- popup 3件、dashboard 7件、background 4件、utils 1件の計15ファイルを移行
- getAllowedUrls は urlWhitelist に、API_KEY_FIELDS/migrateToSingleSettingsObject は settingsMigration に直接 import
- これで src/ の直接的な settingsStore.js 依存は barrel のみになり、二重真実の主要経路を解消
