+++ src/dashboard/encryptedBackupService.ts
@@ -3,7 +3,8 @@
  * 履歴(SQLite DB全体)と設定を1つの暗号化ファイルにまとめてエクスポート/インポートする。
  */
 
-import { getSettings, saveSettings } from '../utils/storage/settingsStore.js';
+import { settingsRepository } from '../utils/storage/SettingsRepository.js';
+import { saveSettings } from '../utils/storage/settingsStore.legacy.js';
 import { exportDb } from './exportLogsService.js';
 import { restoreDb, isServiceError } from './dashboardSqliteService.js';
 import { encryptEnvelope, decryptEnvelope, isEncryptionEnvelope } from '../utils/crypto/index.js';
@@ -23,7 +24,7 @@ interface BackupPayload {
 }
 
 async function buildBackupPayload(): Promise<BackupPayload> {
-  const settings = await getSettings();
+  const settings = await settingsRepository.getAll();
   const dbBlob = await exportDb();
   const dbBuffer = await dbBlob.arrayBuffer();
   const historyDbBase64 = bytesToBase64(new Uint8Array(dbBuffer));
