+++ src/popup/privatePageDialog.ts
@@ -1,6 +1,7 @@
 import type { PendingSave } from './mainTypes.js';
 import { extractDomain } from '../utils/domainUtils.js';
-import { getSettings, saveSettings } from '../utils/storage/settingsStore.js';
+import { settingsRepository } from '../utils/storage/SettingsRepository.js';
+import { saveSettings } from '../utils/storage/settingsStore.legacy.js';
 import { StorageKeys } from '../utils/storage/types.js';
 import { startAutoCloseTimer } from './autoClose.js';
 import { getMessage } from '../utils/i18n.js';
@@ -100,7 +101,7 @@ document.getElementById('dialog-save-domain')?.addEventListener('click', async (
   if (currentPendingSave) {
     const domain = extractDomain(currentPendingSave.url);
     if (domain) {
-      const settings = await getSettings();
+      const settings = await settingsRepository.getAll();
       const whitelist = settings[StorageKeys.DOMAIN_WHITELIST] || [];
       if (!whitelist.includes(domain)) {
         whitelist.push(domain);
@@ -116,7 +117,7 @@ document.getElementById('dialog-save-path')?.addEventListener('click', async () 
   dialog?.close();
 
   if (currentPendingSave) {
-    const settings = await getSettings();
+    const settings = await settingsRepository.getAll();
     const whitelist = settings[StorageKeys.DOMAIN_WHITELIST] || [];
     if (!whitelist.includes(currentPendingSave.url)) {
       whitelist.push(currentPendingSave.url);
